# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer DB API",
        "description": "Customer database API: authentication with access and refresh tokens, customer management, avatars, CSV import/export and background jobs.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "v1.0.0"
    },
    "paths": {
...
        "/customer/export": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Queue customer export",
                "description": "Queue a CSV export of the filtered customer list and return the background job id.",
                "operationId": "3b4b716f4b49221bc1d54c93074e007a",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "keywords": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "lead",
                                            "prospect",
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Export queued"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
        "/customer/import": {
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Queue customer import",
                "description": "Upload a customer CSV and queue the import job. Requires an admin token.",
                "operationId": "cb07ddaa1fe37fe4917f57a04efaaf71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "CSV with a name column; email, phone, company, status, address and notes are optional",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Import queued"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
        "/customer/export/{id}/download": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Download customer export",
                "description": "Download a completed export as a CSV attachment.",
                "operationId": "7d86d3d551632918ce603e0b3901a2a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    },
                    "404": {
                        "description": "Export not found"
                    },
                    "409": {
                        "description": "Export is not completed yet"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
...
        "/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Job status",
                "description": "Status, progress and result of a background job.",
                "operationId": "4978b6e592618931fe959d86a1a55d9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    },
                    "404": {
                        "description": "Job not found"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "auth_token": {
                "type": "apiKey",
                "description": "JWT issued by the application; send the raw token or 'Bearer <token>'.",
                "name": "Authorization",
                "in": "header"
            },
            "health_token": {
                "type": "apiKey",
                "description": "Token configured through HEALTHCHECK_TOKEN for detailed health diagnostics.",
                "name": "X-Health-Token",
                "in": "header"
            }
        }
    },
...
}